as.lazy {lazyeval}	R Documentation

Convert an object to a lazy expression or lazy dots.

Description

Convert an object to a lazy expression or lazy dots.

Usage

as.lazy(x, env = baseenv())

as.lazy_dots(x, env)

Arguments

x
An R object. Current methods for as.lazy() convert formulas, character vectors, calls and names. Methods for as.lazy_dots() convert lists and character vectors (by calling lapply() with as.lazy().)

env
Environment to use for objects that don't already have associated environment.

Examples

as.lazy(~ x + 1)
as.lazy(quote(x + 1), globalenv())
as.lazy("x + 1", globalenv())

as.lazy_dots(list(~x, y = ~z + 1))
as.lazy_dots(c("a", "b", "c"), globalenv())
as.lazy_dots(~x)
as.lazy_dots(quote(x), globalenv())
as.lazy_dots(quote(f()), globalenv())
as.lazy_dots(lazy(x))

[Package lazyeval version 0.2.2 Index]