call_new {lazyeval}	R Documentation

Create a call by "hand"

Description

Create a call by "hand"

Usage

call_new(f, ..., .args = list())

Arguments

f
Function to call. For make_call, either a string, a symbol or a quoted call. For do_call, a bare function name or call.

..., .args
Arguments to the call either in or out of a list

Examples

# f can either be a string, a symbol or a call
call_new("f", a = 1)
call_new(quote(f), a = 1)
call_new(quote(f()), a = 1)

#' Can supply arguments individually or in a list
call_new(quote(f), a = 1, b = 2)
call_new(quote(f), .args = list(a = 1, b = 2))

[Package lazyeval version 0.2.2 Index]