f_unwrap {lazyeval}	R Documentation

Unwrap a formula

Description

This interpolates values in the formula that are defined in its environment, replacing the environment with its parent.

Usage

f_unwrap(f)

Arguments

f
A formula to unwrap.

Examples

n <- 100
f <- ~ x + n
f_unwrap(f)

[Package lazyeval version 0.2.2 Index]