Interpolate values into an expression.

This is useful if you want to build an expression up from a mixture of constants and variables.

interp(`_obj`, ..., .values)

Arguments

_obj

An object to modify: can be a call, name, formula, lazy, or a string.

..., .values

Either individual name-value pairs, or a list (or environment) of values.

Examples

# Interp works with formulas, lazy objects, quoted calls and strings
interp(~ x + y, x = 10)
#> ~10 + y
#> <environment: 0x5618ec247150>
interp(lazy(x + y), x = 10)
#> <lazy>
#>   expr: 10 + y
#>   env:  <environment: 0x5618ec247150>
interp(quote(x + y), x = 10)
#> 10 + y
interp("x + y", x = 10)
#> [1] "10 + y"

# Use as.name if you have a character string that gives a
# variable name
interp(~ mean(var), var = as.name("mpg"))
#> ~mean(mpg)
#> <environment: 0x5618ec247150>
# or supply the quoted name directly
interp(~ mean(var), var = quote(mpg))
#> ~mean(mpg)
#> <environment: 0x5618ec247150>

# Or a function!
interp(~ f(a, b), f = as.name("+"))
#> ~a + b
#> <environment: 0x5618ec247150>
# Remember every action in R is a function call:
# http://adv-r.had.co.nz/Functions.html#all-calls

# If you've built up a list of values through some other
# mechanism, use .values
interp(~ x + y, .values = list(x = 10))
#> ~10 + y
#> <environment: 0x5618ec247150>

# You can also interpolate variables defined in the current
# environment, but this is a little risky.
y <- 10
interp(~ x + y, .values = environment())
#> ~x + 10
#> <environment: 0x5618ec247150>