Capture expression for later lazy evaluation.

lazy() uses non-standard evaluation to turn promises into lazy objects; lazy_() does standard evaluation and is suitable for programming.

lazy_(expr, env)

lazy(expr, env = parent.frame(), .follow_symbols = TRUE)

Arguments

expr

Expression to capture. For lazy_ must be a name or a call.

env

Environment in which to evaluate expr.

.follow_symbols

If TRUE, the default, follows promises across function calls. See vignette("chained-promises") for details.

Details

Use lazy() like you'd use substitute() to capture an unevaluated promise. Compared to substitute() it also captures the environment associated with the promise, so that you can correctly replay it in the future.

Examples

lazy_(quote(a + x), globalenv())
#> <lazy>
#>   expr: a + x
#>   env:  <environment: R_GlobalEnv>

# Lazy is designed to be used inside a function - you should
# give it the name of a function argument (a promise)
f <- function(x = b - a) {
  lazy(x)
}
f()
#> <lazy>
#>   expr: b - a
#>   env:  <environment: 0x5618ea03a268>
f(a + b / c)
#> <lazy>
#>   expr: a + b/c
#>   env:  <environment: 0x5618ea12f6c8>

# Lazy also works when called from the global environment. This makes
# easy to play with interactively.
lazy(a + b / c)
#> <lazy>
#>   expr: a + b/c
#>   env:  <environment: 0x5618ea12f6c8>

# By default, lazy will climb all the way back to the initial promise
# This is handy if you have if you have nested functions:
g <- function(y) f(y)
h <- function(z) g(z)
f(a + b)
#> <lazy>
#>   expr: a + b
#>   env:  <environment: 0x5618ea12f6c8>
g(a + b)
#> <lazy>
#>   expr: a + b
#>   env:  <environment: 0x5618ea12f6c8>
h(a + b)
#> <lazy>
#>   expr: a + b
#>   env:  <environment: 0x5618ea12f6c8>

# To avoid this behavour, set .follow_symbols = FALSE
# See vignette("chained-promises") for details