ast_ {lazyeval}	R Documentation

Display a call (or expression) as a tree.

Description

ast_ takes a quoted expression; ast does the quoting for you.

Usage

ast_(x, width = getOption("width"))

ast(x)

Arguments

x
Quoted call, list of calls, or expression to display.

width
Display width, defaults to current width as reported by getOption("width").

Examples

ast(f(x, 1, g(), h(i())))
ast(if (TRUE) 3 else 4)
ast(function(a = 1, b = 2) {a + b + 10})
ast(f(x)(y)(z))

ast_(quote(f(x, 1, g(), h(i()))))
ast_(quote(if (TRUE) 3 else 4))
ast_(expression(1, 2, 3))

[Package lazyeval version 0.2.2 Index]