auto_name {lazyeval}	R Documentation

Automatically name all components of a lazy dots.

Description

Any components missing a name will automatically get a name added by looking at the first max_width characters of the deparsed expression.

Usage

auto_name(x, max_width = 40)

Arguments

x
A lazy_dots

max_width
Maximum number of characters to use

Examples

x <- lazy_dots(1 + 2, mean(mpg))
auto_name(x)

auto_name(list(~f, quote(x)))

[Package lazyeval version 0.2.2 Index]