lazy_eval {lazyeval}	R Documentation

Evaluate a lazy expression.

Description

Evaluate a lazy expression.

Usage

lazy_eval(x, data = NULL)

Arguments

x
A lazy object or a formula.

data
Option, a data frame or list in which to preferentially look for variables before using the environment associated with the lazy object.

Examples

f <- function(x) {
  z <- 100
  ~ x + z
}
z <- 10
lazy_eval(f(10))
lazy_eval(f(10), list(x = 100))
lazy_eval(f(10), list(x = 1, z = 1))

lazy_eval(lazy_dots(a = x, b = z), list(x = 10))

[Package lazyeval version 0.2.2 Index]